Match frame-graph viewport copies
Preserve Babylon Lite's double-precision viewport values, integer floor bounds, and matching scissor rectangles for SDL_GPU copy tasks. Tighten scene 145 and 146 parity baselines after eliminating fractional MSAA tile coverage.

Copilot-Session: 856ca28c-6ed2-45f1-8032-9ecd2b2958c5

+++ native/include/bblite/runtime.hpp
@@ -150,10 +150,10 @@ struct GeometryTextureDescription {
 };
 
 struct NormalizedViewport {
-    float x = 0.0f;
-    float y = 0.0f;
-    float width = 1.0f;
-    float height = 1.0f;
+    double x = 0.0;
+    double y = 0.0;
+    double width = 1.0;
+    double height = 1.0;
 };
 
 struct RenderTargetOptions {
